expanded all rosters
starting with the 7th roster on each team look for each player value
then an 80 then the second player value. for example, boston is 1A 80 53
00.

If you look at their 7th roster slot (the 1st hidden one), the first
player is 1A and the second player is 53. If you change that to 00,
you'll unlock the 7th (and therefore all) of the remaining Boston
rosters.

P1 80 P2 00     ->     P1 00 P2 00